+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## 1.1.0-SNAPSHOT
 
+* Service pause/resume improvements:
+  - New, simple, file based key/value service context index for storing
+    service serialized instance
+  - Service.setProcessingStage now atomically serializes service state
+    if processing stage is PAUSED
+  Service.setProcessingStage() now returns ServiceRuntimeContext
+
 ## 1.0.0
 
 * Improve Xenon UI's' node selector to support multi node group use cases.